test: dailySummary 테스트

+++ src/main/java/com/weady/weady/domain/tags/entity/TemperatureTag.java
@@ -15,6 +15,6 @@ public class TemperatureTag {
 
     @Column(length = 50)
     private String name;
-    private Integer minTemperature;
-    private Integer maxTemperature;
+    private Float minTemperature;
+    private Float maxTemperature;
 }
